+++ app/code/community/Ecocode/Profiler/Model/AppDev.php
@@ -48,7 +48,7 @@ class Ecocode_Profiler_Model_AppDev extends Mage_Core_Model_App
      * @param array $cacheInitOptions
      * @return Mage_Core_Model_App
      */
-    protected function _initCache(array $cacheInitOptions = array())
+    protected function _initCache(array $cacheInitOptions = [])
     {
         //its to early to make use of normal rewrites as they are not yet loaded
         //so just set it our self
@@ -101,12 +101,18 @@ class Ecocode_Profiler_Model_AppDev extends Mage_Core_Model_App
      */
     protected function _callObserverMethod($object, $method, $observer)
     {
-        $eventName               = $observer->getEvent()->getName();
+        $eventName = $observer->getEvent()->getName();
+
+        $start  = microtime(true);
+        $return = parent::_callObserverMethod($object, $method, $observer);
+        $stop   = microtime(true);
+
         if (!$observer->getEvent()->getData('debug')) {
             $this->calledListeners[] = [
-                'event_name' => $eventName,
-                'class'      => get_class($object),
-                'method'     => $method
+                'event_name'     => $eventName,
+                'class'          => get_class($object),
+                'method'         => $method,
+                'execution_time' => $stop - $start
             ];
 
             if (!method_exists($object, $method)) {
@@ -121,7 +127,7 @@ class Ecocode_Profiler_Model_AppDev extends Mage_Core_Model_App
             }
         }
 
-        return parent::_callObserverMethod($object, $method, $observer);
+        return $return;
     }
 
     public function getEvents()